+++ src/Database/Create/Tables/SegmentLocation.sql
@@ -4,6 +4,6 @@
     [SegmentId] int NOT NULL CONSTRAINT [FK_SegmentLocation_Segment] REFERENCES [Segment],
 
     [Order] int NOT NULL,
-    [Latitude] decimal(25,20) NOT NULL,
-    [Longitude] decimal(25,20) NOT NULL,
+    [Latitude] float NOT NULL,
+    [Longitude] float NOT NULL,
 )
